Like and Dislike graphics added, and URL extraction enabled
The content.js file now carries out a search across the Home Page and extracts the URLs of the Youtube videos being displayed. These URLs will be later used for calling the Youtube Data API for getting the likes and dislike numbers.

Further, the content.js also identifies the places where the titles of the videos are and places the Like and Dislike images accordingly. Later, the data shall be placed adjacent to these images.

The "assets" directory contains the image files (.jfif), "like" and "dislike", being displayed.