Bugs Removed and UI Features Added
The bug that lead to duplication of like and dislike counts, and graphics has been removed.

The Youtube Data API is no longer used. Instead "returnyoutubedislike.com" API is used for like as well as dislike count. For now, the googleapi.js has been removed. In case its need arises in the future, it shall be added again.

Code processes have been considerably streamlined. Various functions have been created for clear workflow and to ensure sequential execution of functions and events. The async function has been used to achieve this goal.

Several console.log statements have been commented and not discarded to aid debugging in the future. These statements are at critical places in the code. Once uncommented, they can be monitored to check for failures in the code's workflow.

UI features:
The like and dislike count have been added next to the like and dislike icons. Using the like percentage, the Youtube videos are marked as "green" and "recommended" if the like percentage is greater than 98%.

If the like percentage is less than 40%, then the video is marked as "red".

If there are no likes and dislikes at all then the video is marked as "grey".

Rest of the videos are not marked with any colors.

These features enable the users to take quick judgements as to whether the video is worth their time or not.